IDM-2090: Make sure message IDs are unique
The sleep makes sure that the message ID will be unique on the
next invocation, Synchronized makes sure there are no concurrent
invocations. Millisecond time is used to guarantee that the perl
daemon will pick them in the right order.

As far as I can tell this is good code but to be honest I have
no idea what I am doing here.